Expose API methods for explicit open and close operations
This is useful for creating idempotent operations. For my particular use
case, Home Assistant's local polling of the OpenGarage device leads to
situations where HA gets out of sync with the physical state of the
garage door, and automation rules can wrongly trigger the door to open
or close.

Allowing HA to assert an *intent* via explicit open or close instead of
a simple "toggle" means fewer scratches on my car's roof from the garage
door closing on it.